+++ internal/virt/utils_test.go
@@ -1,0 +1,23 @@
+package virt
+
+import (
+	"bytes"
+	"io"
+	"os"
+)
+
+// Helper function to capture stdout
+func captureStdout(f func()) string {
+	old := os.Stdout
+	r, w, _ := os.Pipe()
+	os.Stdout = w
+
+	f()
+
+	w.Close()
+	os.Stdout = old
+
+	var buf bytes.Buffer
+	io.Copy(&buf, r)
+	return buf.String()
+}
